商品削除機能(修正)

+++ app/controllers/items_controller.rb
@@ -17,6 +17,8 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+    return unless @item && @item.user == current_user
+
     if @item.destroy
       redirect_to root_path
     else
